Organize String Algorithms
