+++ Algorithms.Tests/Strings/GeneralStringAlgorithmsTests.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using Algorithms.Strings;
+using Algorithms.Strings.PatternMatching;
 using NUnit.Framework;
 
 namespace Algorithms.Tests.Strings
