+++ Algorithms.Tests/Strings/KnuthMorrisPrattSearcherTests.cs
@@ -1,4 +1,5 @@
 ﻿using Algorithms.Strings;
+using Algorithms.Strings.PatternMatching;
 using NUnit.Framework;
 
 namespace Algorithms.Tests.Strings
