+++ Algorithms.Tests/Strings/RabinKarpTests.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using Algorithms.Strings;
+using Algorithms.Strings.PatternMatching;
 using NUnit.Framework;
 
 namespace Algorithms.Tests.Strings
