+++ Algorithms/Strings/PatternMatching/GeneralStringAlgorithms.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Algorithms.Strings
+namespace Algorithms.Strings.PatternMatching
 {
     /// <summary>
     ///     Implements simple algorithms on strings.
