+++ Algorithms/Strings/PatternMatching/KnuthMorrisPrattSearcher.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace Algorithms.Strings
+namespace Algorithms.Strings.PatternMatching
 {
     public class KnuthMorrisPrattSearcher
     {
